Typescriptify alignments track model, reduce SNPCoverage subtrack height, reduce ticks when smaller
