+++ packages/alignments/src/AlignmentsTrack/model.ts
@@ -1,0 +1,86 @@
+import {
+  getConf,
+  ConfigurationReference,
+} from '@gmod/jbrowse-core/configuration'
+import { BaseTrack } from '@gmod/jbrowse-plugin-linear-genome-view'
+import { types, addDisposer } from 'mobx-state-tree'
+import { autorun } from 'mobx'
+import AlignmentsTrackComponent from './components/AlignmentsTrack'
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export default (pluginManager: any, configSchema: any) => {
+  return types
+    .compose(
+      'AlignmentsTrack',
+      BaseTrack,
+      types
+        .model({
+          PileupTrack: types.maybe(
+            pluginManager.getTrackType('PileupTrack').stateModel,
+          ),
+          SNPCoverageTrack: types.maybe(
+            pluginManager.getTrackType('SNPCoverageTrack').stateModel,
+          ),
+          type: types.literal('AlignmentsTrack'),
+          configuration: ConfigurationReference(configSchema),
+          height: 250,
+        })
+        .volatile(() => ({
+          ReactComponent: AlignmentsTrackComponent,
+        })),
+    )
+    .views(self => ({
+      get pileupTrackConfig() {
+        return {
+          ...getConf(self),
+          type: 'PileupTrack',
+          name: `${getConf(self, 'name')} pileup`,
+          trackId: `${self.configuration.trackId}_pileup`,
+        }
+      },
+
+      get layoutFeatures() {
+        return self.PileupTrack.layoutFeatures
+      },
+
+      get snpCoverageTrackConfig() {
+        return {
+          ...getConf(self),
+          type: 'SNPCoverageTrack',
+          name: `${getConf(self, 'name')} snpcoverage`,
+          trackId: `${self.configuration.trackId}_snpcoverage`,
+          height: 10,
+          adapter: {
+            type: 'SNPCoverageAdapter',
+            subadapter: getConf(self, 'adapter'),
+          },
+        }
+      },
+    }))
+    .actions(self => ({
+      afterAttach() {
+        addDisposer(
+          self,
+          autorun(() => {
+            this.setSNPCoverageTrack(self.snpCoverageTrackConfig)
+            this.setPileupTrack(self.pileupTrackConfig)
+          }),
+        )
+      },
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      setSNPCoverageTrack(trackConfig: any) {
+        self.SNPCoverageTrack = {
+          type: 'SNPCoverageTrack',
+          configuration: trackConfig,
+        }
+        self.SNPCoverageTrack.height = 40
+      },
+      // eslint-disable-next-line @typescript-eslint/no-explicit-any
+      setPileupTrack(trackConfig: any) {
+        self.PileupTrack = {
+          type: 'PileupTrack',
+          configuration: trackConfig,
+        }
+      },
+    }))
+}
