+++ packages/wiggle/src/WiggleTrack/components/WiggleTrackComponent.tsx
@@ -21,11 +21,13 @@ export const YScaleBar = observer(
       range: [height, 0],
       inverted: getConf(model, 'inverted'),
     })
-    const axisProps = axisPropsFromTickScale(scale, 4)
+    const ticks = height < 50 ? 2 : 4
+    const axisProps = axisPropsFromTickScale(scale, ticks)
     const values =
       scaleType === 'log'
         ? axisProps.values.filter((s: number) => powersOfTen.includes(s))
         : axisProps.values
+
     return (
       <svg
         style={{
